requirements.txt
textual
pymongo
boto3
python-dotenv
textual-serve
